+++ CHANGELOG.md
@@ -5,6 +5,13 @@
   - Still flagging non-custom-error second arguments (e.g., string literals, numbers, variables)
   - Still flagging `require` statements with no second argument
 
+## [4.0.1] - XXXX-XX-XX
+
+### Fixed
+- false positive in no-unused-var when function parameters are only used only as
+  modifier arguments
+  https://github.com/solhint-community/solhint-community/pull/159
+
 ## [4.0.0] - 2024-04-10
 Stable release.
 Current `latest` version.